fixed install script

+++ install.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 
+git pull
 
 REPO_BASE="$(dirname "$PWD/$0")"
 REPO_BASE="$(readlink -f $REPO_BASE)"
@@ -7,8 +8,10 @@ TRASH=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
 TRASH="/tmp/old.emacs/$TRASH"
 mkdir -p $TRASH
 
-[ -e ~/.emacs ] && mv ~/.emacs $TRASH
-ln -s $REPO_BASE/.emacs ~/.emacs
+if [ -e ~/.emacs -a ! -L ~/.emacs ]; then
+    mv ~/.emacs $TRASH
+    ln -s $REPO_BASE/.emacs ~/.emacs
+fi
 
 [ ! -d ~/.emacs.d ] && mkdir ~/.emacs.d/
